Modified structure of package. Left a simpler __init__ file. Fixes #4.
